area_api/requirements.txt
aniso8601==8.1.1
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
schematics==2.1.0
six==1.15.0
tzlocal==2.1
urllib3==1.26.3
Werkzeug==1.0.1
PyGithub==1.54.1
pytest==6.2.2
